Added welcome page
Added a welcome page that shows the features that are currently implemented. A servlet has been added in the control package that is opened on the default URL. The servlet forwards to a JSP that has two buttons that redirect to the other 2 JSPs (although it should redirect to otehr servlets...). A command has been added in the sql creation script to add theclient user