GAME UPLOAD: added game upload page
You can now access the admin-restricted game upload page and upload a game. The pegi and status values using the html select input type. The game is implemented server-side using the DAO/DTO pattern. In the Game DTO, the date is implemented with a simple string and the State and Pegi values are implemented using Enums.