Added footer to all pages
The footer works by putting the page's main content in a 'main' section. The body is display;flex and main is the only element of the flexbox, taking up all the space in the page. Then the header and footer are set with a min-height which allows them to go exactly to the top and to the bottom of the page somehow.