ADDED: edit game, cart total price. FIXED: add to wishlist not working, minor regex bug, login error message color and position, wishlist button does not show up if game is unlisted
You can edit the game as admin in a page that is similar to the game upload page. The categories, requirements and images are first deleted and then the new ones are inserted. Images are only updated if new ones are inserted. Total prices are calculated again without any particular regard for performance. The original images and requirements tables are not deleted, only the table that unites them with the game